server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.2
openai==1.51.0
httpx==0.27.2
langchain==0.3.1
langchain-community==0.3.1
pypdf==4.3.1
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1